+++ R/cohort_methods.R
@@ -483,94 +483,133 @@ Cohort <- R6::R6Class(
     },
     #' @description
     #' Return reproducible data filtering code.
-    #' @param step_id If `step_id` specified and `filter_id` missing, reproducible
-    #'     code for specified step is returned.
-    #'     If `filter_id` not missing, `step_id` defines from which step the filter should be taken.
-    #' @param filter_id If not missing, precises the filter for which reproducible code should be returned.
+    #' @param include_source If `TRUE` source generating code will be included.
+    #' @param include_methods Which methods definition should be included in the result.
+    #' @param include_action Which action should be returned in the result.
+    #'     `pre_filtering`/`.post_filtering` - to include data transformation before/after filtering.
+    #'     s`run_binding` - data binding transformation.
+    #' @param modifier A function taking data frame (storing reproducible code metadata) as
+    #'     an argument, and returning data frame with `expr` column which is then
+    #'     combined into a single expression (final result of `get_code`).
+    #'     See \link{.repro_code_tweak}.
+    #' @param mark_step Include information which filtering step is performed.
     #' @param ... Other parameters passed to \link[formatR]{tidy_source}.
-    get_code = function(step_id, filter_id, ...) {
-      source_expr <- if (!is.null(private$source$source_code)) {
-        private$source$source_code
-      } else {
-        substitute(source <- list(dtconn = x), list(x = attr(private$source$dtconn, "call")))
-      }
-      source_type <- class(private$source)[1]
+    get_code = function(
+      include_source = TRUE, include_methods = c(".pre_filtering", ".post_filtering", ".run_binding"),
+      include_action = c("pre_filtering", "post_filtering", "run_binding"),
+      modifier = .repro_code_tweak, mark_step = TRUE, ...) {
 
-      init_step_expr <- parse_func_expr(
-        .get_method(paste0(".init_step", ".", source_type))
-      )
-      run_binding_fun_expr <- quote({})
-      bind_keys_expr <- quote({})
-      pre_run_binding_expr <- quote({})
-      run_binding_expr <- quote({})
-      pre_filtering_expr <- post_filtering_expr <- quote({})
-      pre_filtering_fun_expr <- method_to_expr(".pre_filtering", source_type)
-      post_filtering_fun_expr <- method_to_expr(".post_filtering", source_type)
-      if (length(pre_filtering_fun_expr) > 1) {
-        pre_filtering_expr <- quote({
-          data_object <- .pre_filtering(source, data_object, step_id)
-        })
-      }
-      if (length(post_filtering_fun_expr) > 1) {
-        post_filtering_expr <- quote({
-          data_object <- .post_filtering(source, data_object, step_id)
-        })
-      }
-      if (!is.null(private$source$binding_keys)) {
-        run_binding_fun_expr <- method_to_expr(".run_binding", source_type)
-        if (isTRUE(private$source$options$display_binding)) {
-          bind_keys_expr <- assign_expr(quote(binding_keys), private$source$binding_keys)
-        }
-        pre_run_binding_expr <- quote({
-          pre_data_object <- data_object
-        })
-        run_binding_expr <- quote({
-          for (binding_key in binding_keys) {
-            data_object <- .run_binding(
-              source, binding_key,
-              pre_data_object,
-              data_object
-            )
+      source_type <- class(private$source)[1]
+      # todo improve
+      fun_args <- environment()
+      code_params <- c(
+        "include_source", "include_methods", "include_action", "modifier", "mark_step"
+      ) %>%
+        stats::setNames(nm = .) %>%
+        purrr::map(
+          ~if (is.null(self$attributes[[.x]]) & !.x %in% names(self$attributes)) {
+            fun_args[[.x]]
+          } else {
+            self$attributes[[.x]]
           }
-        })
-      }
-      if (!missing(filter_id)) {
-        exprs_list <- parse_filter_expr(private$steps[[step_id]]$filters[[filter_id]])
-        return(combine_expressions(exprs_list))
-      }
-      get_step_code <- function(step_id, private) {
-        filtering_expr <- private$steps[[step_id]]$filters %>%
-          purrr::keep(~ .x$get_params("active")) %>%
-          purrr::map(~ parse_filter_expr(.x))
-        list(
-          glue::glue("# step {step_id} "),
-          substitute({step_id <- step_value}, list(step_value = step_id)),
-          pre_filtering_expr,
-          filtering_expr,
-          post_filtering_expr,
-          pre_run_binding_expr
+        )
+
+      code_components <- list()
+
+      for (extra_method in code_params$include_methods) {
+        code_components <- append(
+          code_components,
+          type_expr(type = "meta", expr = method_to_expr(extra_method, source_type))
         )
       }
-      if (!missing(step_id)) {
-        exprs_list <- get_step_code(step_id, private)
-      } else {
-        exprs_list <- names(self$get_step()) %>%
-          purrr::map(get_step_code, private = private)
+      if (code_params$include_source) {
+        code_components <- append(
+          code_components,
+          type_expr(type = "source", expr = get_source_expr(source_type, self, private))
+        )
+      }
+      for (step_id in names(self$get_step())) {
+        if (code_params$mark_step) {
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "step_init", step = step_id,
+              expr = rlang::expr(step_id <- !!step_id)
+            )
+          )
+        }
+        # todo add binding keys
+        if ("run_binding" %in% code_params$include_action) {
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "run_binding", step = step_id,
+              expr = rlang::expr(
+                pre_data_object <- data_object
+              )
+            )
+          )
+        }
+        if ("pre_filtering" %in% code_params$include_action) {
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "pre_filtering", step = step_id,
+              expr = rlang::expr(
+                data_object <- .pre_filtering(source, data_object, !!step_id)
+              )
+            )
+          )
+        }
+        active_filters <- private$steps[[step_id]]$filters %>%
+          purrr::keep(~ .x$get_params("active"))
+        for (filter in active_filters) {
+          filter_params <- filter$get_params()
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "filtering", step = step_id,
+              expr = parse_filter_expr(filter),
+              !!!filter_params
+            )
+          )
+        }
+        if ("post_filtering" %in% code_params$include_action) {
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "post_filtering", step = step_id,
+              expr = rlang::expr(
+                data_object <- .post_filtering(source, data_object, !!step_id)
+              )
+            )
+          )
+        }
+        if ("run_binding" %in% code_params$include_action) {
+          code_components <- append(
+            code_components,
+            type_expr(
+              type = "run_binding", step = step_id,
+              expr = rlang::expr(
+                for (binding_key in binding_keys) {
+                  data_object <- .run_binding(
+                    source, binding_key,
+                    pre_data_object, data_object
+                  )
+                }
+              )
+            )
+          )
+        }
       }
 
-      exprs_list <- append(
-        append(
-          list(
-            bind_keys_expr, run_binding_fun_expr, pre_filtering_fun_expr, post_filtering_fun_expr,
-            source_expr, init_step_expr
-          ),
-          exprs_list
-        ),
-        list(run_binding_expr, quote({data_object}))
-      )
+      code_components_df <- code_components %>%
+        purrr::map_dfr(function(x) x) %>%
+        dplyr::filter(purrr::map_lgl(expr, ~!is.null(.)))
 
-      # todo code include once
-      res_quote <- combine_expressions(unlist(exprs_list))
+      code_components_df <- code_params$modifier(private$source, code_components_df)
+      # todo code include once?
+      res_quote <- combine_expressions(unlist(code_components_df$expr))
       formatR::tidy_source(
         text = as.character(res_quote)[-1],
         ...
@@ -608,11 +647,11 @@ Cohort <- R6::R6Class(
         data_object = private$data_objects[[prev_step(step_id)]],
         step_id = step_id
       )
-      for (data_filter in self$get_filter(step_id)) {
-        if (data_filter$get_params("active")) {
-          temp_data_object <- temp_data_object %>%
-            data_filter$filter_data()
-        }
+      active_filters <- self$list_active_filters(step_id)
+      for (filter_id in active_filters) {
+        data_filter <- self$get_filter(step_id, filter_id)
+        temp_data_object <- temp_data_object %>%
+          data_filter$filter_data()
       }
 
       private$data_objects[[step_id]] <- .post_filtering(
@@ -641,6 +680,8 @@ Cohort <- R6::R6Class(
         if (!is_cached) {
           self$update_cache(step_id, filter_id, state = "pre")
         }
+      }
+      for (filter_id in active_filters) {
         self$update_cache(step_id, filter_id, state = "post")
       }
 
@@ -1075,17 +1116,27 @@ stat <- function(x, step_id, filter_id, ..., state = "post") {
 #' Return reproducible data filtering code.
 #'
 #' @param x Cohort object.
-#' @param step_id If `step_id` specified and `filter_id` missing, reproducible
-#'     code for specified step is returned.
-#'     If `filter_id` not missing, `step_id` defines from which step the filter should be taken.
-#' @param filter_id If not missing, precises the filter for which reproducible code should be returned.
+#' @param include_source If `TRUE` source generating code will be included.
+#' @param include_methods Which methods definition should be included in the result.
+#' @param include_action Which action should be returned in the result.
+#'     `pre_filtering`/`.post_filtering` - to include data transformation before/after filtering.
+#'     s`run_binding` - data binding transformation.
+#' @param modifier A function taking data frame (storing reproducible code metadata) as
+#'     an argument, and returning data frame with `expr` column which is then
+#'     combined into a single expression (final result of `get_code`).
+#'     See \link{.repro_code_tweak}.
+#' @param mark_step Include information which filtering step is performed.
 #' @param ... Other parameters passed to \link[formatR]{tidy_source}.
 #' @return \link[formatR]{tidy_source} output storing reproducible code for generating final step data.
 #'
 #' @seealso \link{cohort-methods}
 #' @export
-code <- function(x, step_id, filter_id, ...) {
-  x$get_code(step_id, filter_id, ...)
+code <- function(x, include_source = TRUE, include_methods = c(".pre_filtering", ".post_filtering", ".run_binding"),
+                 include_action = c("pre_filtering", "post_filtering", "run_binding"),
+                 modifier = .repro_code_tweak, mark_step = TRUE, ...) {
+  x$get_code(include_source = include_source, include_methods = include_methods,
+             include_action = include_action,
+             modifier = modifier, mark_step = mark_step, ...)
 }
 
 #' Get step related data