+++ NEWS.md
@@ -1,3 +1,9 @@
+# cohortBuilder 0.1.2
+
+* Changed the way reproducible code is returned. Now more flexibility is allowed with using e.g. `.repro_code_tweak` method.
+* The `tblist` source reproducible code is now using pipe chains for each dataset filtering.
+* Optimized filtering with having cache computed only for active filters.
+
 # cohortBuilder 0.1.1
 
 * Add `.post_binding` method, that allows to modify data object when binding is completed.